Add a header for the f142 module, step towards testability
